feat(collection): move add-by-URL into SuggestedTracks panel; remove confusing IconLink
Ray flagged the copy-link-looking IconButton on the playlist owner header
as confusing — its actual behavior (open a "paste track URLs" modal) wasn't
discoverable from the icon, and adding tracks by URL belongs in the same
section users already think of as "where I add tracks."

Per Julian's UX suggestion, the natural action for "add by URL" is to
paste — no separate modal or icon required. Moves the feature inline into
the SuggestedTracks panel (already titled "Add some tracks", already the
obvious place to add) as an always-visible TextInput that paste-detects.

Changes:
- packages/web/src/components/suggested-tracks/SuggestedTracks.tsx: add an
  always-visible TextInput row between the heading and the collapsible
  suggestion list. onPaste: if the clipboard contains at least one valid
  Audius track URL (parsed via getPathFromTrackUrl), prevent default,
  resolve via sdk.tracks.getBulkTracks, run the same dedupe + capacity +
  delay-spaced dispatch the modal did, toast a summary. onKeyDown(Enter)
  handles a typed URL too. If the pasted text isn't a URL the input
  behaves like a normal text field — search-field UX. Routes adds through
  draftCollectionCache when the collection is unsaved (mirrors the
  suggested-row Add button's draft branch).
- packages/web/src/components/collection/desktop/OwnerActionButtons.tsx:
  drop the IconLink IconButton and its modal-opener handler / messages /
  hook usage. Header is now lean (Edit, Share, Publish, OverflowMenu).
- Delete the now-unused AddTracksByUrlModal entirely: the React component
  (packages/web/src/components/add-tracks-by-url-modal/), the common
  store slice (packages/common/src/store/ui/modals/add-tracks-by-url-modal/),
  and its registrations in parentSlice, reducers, types, the modals
  re-export, and Modals.tsx. No other call sites.

Feature parity preserved:
- Same parser (one URL per line, or comma/tab separated) — multi-URL paste
  still works.
- Same dedupe vs existing playlist contents.
- Same PLAYLIST_TRACK_LIMIT (100) capacity check.
- Same getBulkTracks SDK call + userTrackMetadataFromSDK transform.
- Same summary toast format (added • duplicates • not found • invalid •
  over-limit).
- Same playlist-only / non-DDEX surface (SuggestedTracks already gated on
  isOwner && !isAlbum && playlist, mirrors the old IconLink condition).

Verification:
- tsc --noEmit clean in packages/common and packages/web.
- Visual smoke needs the full protocol stack (audius web requires Docker
  per CLAUDE.md) — flagging on the PR for manual review.
