+++ packages/common/src/store/ui/modals/index.ts
@@ -43,4 +43,3 @@ export * from './coin-success-modal'
 export * from './fan-club-details-modal'
 export * from './create-playlist-modal'
 export * from './duplicate-playlist-modal'
-export * from './add-tracks-by-url-modal'
