+++ packages/common/src/store/ui/modals/parentSlice.ts
@@ -86,8 +86,7 @@ export const initialState: BasicModalsState = {
   VerificationSuccess: { isOpen: false },
   VerificationError: { isOpen: false },
   CreatePlaylistModal: { isOpen: false },
-  DuplicatePlaylistModal: { isOpen: false },
-  AddTracksByUrlModal: { isOpen: false }
+  DuplicatePlaylistModal: { isOpen: false }
 }
 
 const slice = createSlice({
