+++ packages/common/src/store/ui/modals/types.ts
@@ -122,7 +122,6 @@ export type Modals =
   | 'VerificationError'
   | 'CreatePlaylistModal'
   | 'DuplicatePlaylistModal'
-  | 'AddTracksByUrlModal'
 
 export type BasicModalsState = {
   [modal in Modals]: BaseModalState
